requirements.txt
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.33.9
botocore==1.33.9
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-guardian==2.4.0
django-smart-selects==1.6.0
django-storages==1.14.2
django-tenants==3.5.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
hikvisionapi==0.3.2
httpcore==1.0.2
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
oauthlib==3.2.2
ordered-set==4.1.0
paho-mqtt==1.6.1
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==1.26.16
xmltodict==0.13.0
zope.interface==6.1
